Add files via upload
The program will accept a string value S as the input. The program will print the characters which are present at the odd positions if the count of vowels in S is odd. Else the program will print the characters which are present at the even positions as the output. If there is no vowel in the string S then the program will print -1 as the output