Create string
The program will accept two string values S1 and S2 as the input. The program will print the first character in S1, then the program will print the second character in S2, then the program will print the third character in S1, then the program will print the fourth character in S2 and so on.