Rename matrix to matrices
The program will accept two integer matrices are of the same size RxC as the input. If both the matrices have the same elements in the same positions then retain the elements in their same positions of the first matrix. Else the program will replace the elements by * in the first matrix. Finally, the program will print the modified first matrix as the output.