Create odd digits reverse pattern
The program will accept an integer N as the input. The program will print the odd digits from the last digit and remove the last digit of N then again print the odd digits from the last digit and remove the last digit and so on. If there is no odd digit in the original N then print -1 as the output.