Create Sum of First K Integers
The program will accept N integers and an integer K as the input. The program will find the sum of the first K integers as S. The program will consider the last K integers and for each integer X among the last K integers, the program will print difference between S and X as the output.