Create Maximum Negative Elements - Column
The program will accept an integer matrix of size RxC as the input. The program will print the column number which has the most number of negative numbers in the given matrix as the output. If more than one column have the most number of negative numbers then the program must print the first ocurring such column number as the output. If the matrix does not contain any negative element then the program will print -1 as the output.