Create Change as Multiples of X
The program will accept N integers and an integer X as the input. For each multiple of X among the N integers, the program will change all the values following it to that multiple of X, until encountering another multiple of X. For the last multiple of X, the program will change all the remaining values to the last multiple of X. Finally, the program will print the modified N integers as the output.