Create Repeat First X and Last X
The program will accept a string S and an integer X as the input. The program will form a new string by concatenating the first X and the last X characters of S alternatively. The length of the new string will be equal to the length of S. Finally, the program will print the new string as the output.