Create Divide by Unit Digit
The program will accept N integers as the input. The program will divide each integer by its unit digit and print the result up to two decimal places as the input. If the unit digit of the integer is 0 then the program will print the same integer with the precision up to two decimal places as the output.