Create Spilt Array - Equal Sum
The program will accept an integer array of size N as the input. The program will print YES if it is possible to split the array into two sides so that the sum of the integers on one side is equal to the sum of the integers on the other side. Else the program will print NO as the output.