Create CopyAll and Paste Operations
The program will accept an integer N and a string S as the input. The program will perform two operations they are copyAll and paste repeatedly for N times. Each time when the copyAll operation is performed, all the characters in the string S are copied. The paste operation pastes the last copied characters at the end of the string S. Finally, the program will print the number of characters present in the string S after repeating the two operations for N times.