Create Maximum Number of 1's - Row
The program will accept a matrix of size NxN as the input, the matrix contains only 0's and 1's. The program will count the number of 1's in each row. Then the program will print the row number of the row containing the maximum number of 1's as the output. If two or more rows contain the maximum number of 1's then print the first occuring row number in the given order as the output.