Create Maximum Integer
The program will accept N integers and an integer K as the input. The program will find the maximum integer among N integers as X. Then the program will print X if the value of all other integer values when multiplied by K is less than or equal to X. Else the program will print -1 as the output.