Create Odd Strength
The program will accept N integers as the input. The program will print the output based on the conditions given below. - If the even integers have occurred more number of times than the odd integers then the program will print even as the output. - If the odd integers have occurred more number of times than the even integers then the program will print odd as the output. - If the even integers and the odd integers have occurred equal number of times then the program will print -1 as the output.