Create Swap the Vowels
The program will accept two string values S1 and S2 containing only alphabets. The number of vowels in S1 will always be equal to the number of vowels in S2. The program will replace the vowels in S2 with the vowels in S1 in the same order of occurrence and vice versa. Finally, the program will print the modified string values S1 and S2 as the output.