Create Binary - N Integers
The program will accept N integers as the input. The program will form an integer K using N integers based on the following condition. The binary representation of the integer K must be formed by using the second bit from the end (last but one bit) of every integer among the N integers in the given order. Finally the program will print the integer K as the output.