Create Swap - Multiples of X and Y
The program will accept N integers and two integers X and Y as the input. The program will swap the first occurring mutiple of X and the last occurring multiple of Y among the N integers. Then the program will print modified values of the N integers as the output.