Create Max Possible Sum - Multiples of 10
The program will accept N integers as the input. The program will print the integer X from the given N integers so that the maximum possible sum of remaining integers (except the integer X) is a multiple of 10. If there is no such integer to print, the program will print -1 as the output.