Create String - Rotate K Times
The program will accept two string values S1 and S2 and an integer K representing the number of rotation as the input. The program will print YES if the string S2 is formed from S1 by rotating the string S1 for K times in either clockwise or anti-clockwise direction. Else the program will print NO as the output.