Create Minimum Sum - Left or Right
The program will accept a list of unique positive integers and a positive integer X as the input. The program will find the sum of integers left to integer X then the program will find the sum of integers right to integer X. The program will print the minimum sum value among the obtained two sum values as the output. If there is no integer before or after integer X, the program will print 0 as the minimum sum value.