Create Form Integers - Unit Digit & Tenth Digit
The program will accept N integers as the input. For every two consecutive integers X and Y among the N integers, the program will form an integer by concatenating the unit digit of X and the tenth digit of Y. Then the program will print all the obtained integers in the order of occurrence as the output.