Create Burgers - Maximum Possible Amount
In a restaurant, two types of burgers are available (Veg burgers and Chicken burgers). To assemble a veg burger, 2 buns and a veg cutlet are required. To assemble a chicken burger, 2 buns and a chicken cutlet are required. There are B buns, V veg cutlets and C chicken cutlets are available to assemble the burgers. The cost of one veg burger is Rs. X and the cost of one chicken burger is Rs. Y. The values of B, V, C, X and Y are passed as the input to the program. The program will print the maximum possible amount obtained by selling the burgers as the output.