Create First and Last Rows or Columns
The program will accept an integer matrix of size NxN as the input. If N is odd then the program will print the integers in the first and last rows of the matrix as the output. Else the program will print the integers in the first and last columns of the matrix as the output.