Create Form One and Zero
The program will accept a string S containing only z, e, r, o and n as the input. The program will rearrange the string S such that the modified string containing only "one" and "zero" (all one's followed by all zero's). If it is not possible to rearrange the string S then the program will print -1 as the output.