Create Divide Equally
The program will accept four positive integers as the input. The program will divide the four integers into two parts and print YES if the sum of integers in the first part is equal to the sum of integers in the second part. Else the program will print NO as the output