Create Minimum Possible Integer
The program will accept two integers n and k as the input. the program will print the minimum possible integer x which can be obtained by modifying at most k digits in n as the output. note: the number of digits in n and x should be equal without any leading zeros.