Create Strictly Increasing Matrix
The program will accept an integer matrix of size RxC as the input. The program will print YES if the matrix is a strictly increasing matrix. Else the program will print NO as the output. In a strictly increasing matrix, all the integers in each row (left to right) and all the integers in each column (top to bottom) are in strictly increasing order.