Create Pad with Asterisks - RightPad with Asterisks - RightLeft
Two string values S1 and S2 are passed as input to the program. The program will pad S1 with asterisks on the right (when the length of S2 is greater than the length of S1) and will pad S2 with asterisks on the left (when the length of S1 is greater than the length of S2).