Create 2 Power K + 1 or 2 Power K - 1<
Given a positive integer N as an input, the program must check if N can be expressed as 2^K+1 or 2^K-1 (where K is a positive integer) and print K as the output. If N cannot be expressed as 2^K+1 or 2^K-1, the program must print No as the output.