Create Products & Advertisements
In a super market, there is a list consisting of N products which are numbered from 0 to N - 1. The names of the products are passed as the input to the program. The super market has advertisements to attract customers. The program also accepts a list of integers representing the product number which is advertised as the input. The program must print the product name having the maximum number of advertisements as the output. If more than one product is having the maximum number of advertisements then the program must print the first occurring product as the output.