Create Concatenate T Ones
The program must accept an integer N as the input. The program must find the number of ones T in the binary representation of N. Then the program must form the binary representation of an integer X by concatenating T ones. Finally, the program must print the integer value of X as the output.