Create wo Arrays - Larger from End
The program must accept two integer arrays of size M and N (where M != N). Then it must compare the elements of both the arrays at the same position starting from the end and print the larger value. When M is not equal to N, M-N elements in the beginning of the M sized array will not have a pair and hence will be printed as such.