Create Swap Characters - Pairs
he program accept a string S and N pairs as the input. Each pair contains two integers P1 and P2 representing the position of two characters in the string S. For each pair, the program must swap the characters in P1 and P2. Finally, the program must print the modified string S as the output