Create Sum - Right Side
The program must accept N integers as the input. For each integer X among the N integers, the program must print the sum of all the integers which are less than X and right to X as the output. If there is no such integer on the right side of X then the program must print 0 as the output.