Create Sum - 2^X Integers from Xth Integer
The program must accept N integers and an integer X as the input. The program must print the sum of 2^X integers from the X th integer as the output. If there is no 2^X integers from the Xth integer then print the sum of all integers from the X th  integer as the output.