Create Grapes Distribution
In a box, there are  A  green grapes,  B  purple grapes and  C black grapes. There are three boys B1, B2 and B3 who are all grape lovers. The boys B1, B2 and B3 should eat at least  X  ,  Y  and  Z  grapes respectively with the following conditions. 
- B1 eats only green grapes. 
- B2 never eats black grapes. 
- B3 eats any kind of grapes. 
The values of X, Y, Z, A, B and C are passed as input to the program. The program must print YES  if it is possible to distribute the grapes so that everyone gets their respective number of grapes (at least X, Y and Z). Else the program must print  NO  as the output.