Create String Decryption - Increasing Order
The program must accept a string S (encrypted string) containing only alphabets as the input. The encryption algorithm is given below.       The first character occurs only once.
The second character occurs twice. 
The third character occurs thrice and so on
The program must decrypt the string S and print it as the output. 
Note: The string S is always a valid encrypted string.