Create Jump - Left Right
A boy is currently at the point 0 on a coordinate axis. He jumps by the following rules. The first jump is X units to the right, the second jump is Y units to the left, the third jump is X units to the right, the fourth jump is Y units to the left and so on.
he program must accept three integers X, Y and K as the input. The program must print the position of the boy after K jumps as the output.