Create Sum - Inverted Powers of Two
The program must accept an integer N as the input. The program must generate a sequence of integers containing the integers from 1 to N. Then the program must invert the sign of all the powers of 2 in the generated sequence. Finally, the program must print the sum of all the integers in the modified sequence as the output.