Create Prepare Fruit Salad
There are X kgs of Apple, Y kgs of Banana and Z kgs of Watermelon in a fruit shop. To prepare a box of fruit salad (7kgs), the fruits are required in the ratio of 1 : 2 : 4 (Apple : Banana : Watermelon). The program must print the maximum number of boxes of fruit salad that can be prepared from the given quantity of fruits as the output.