Create Three Tanks with Equal Water
There are three water tanks A, B and C are filled with X, Y and Z litres of water respectively. The values of X, Y and Z are passed as the input to the program. The program must print the minimum quantity of additional water (in litres) is required to fill all three tanks with equal water.
Note: The water can not be drained from any tank.