Create Arrays - Distinct Elements Count
Two integer arrays of length L1 and L2 are passed as input. The program must print the sum of the distinct elements present in both the arrays.
That is the program must print the count of integer values that are present in the first array but not in the second array plus the count of integer values that are present in the second array but not in the first array.