Create Strict Professor - Class Cancellation
In a college, a professor is strict and mandates that at least M students out of the total N students must arrive on time to his class. Else he would cancel the class. The start time of the class is also passed as the input. The individual arrival time for N students is also passed as input in 24 hour format.

The program must print if the class was cancelled or not.