Create Buy X Get Y Free
In a supermarket, there is an offer for chocolates to attract customers. The offer is "BUY X GET Y FREE" (i.e., If a customer buys X chocolates then he/she takes Y chocolates for free). This special offer can be used any number of times. The cost of each chocolate is M rupees. A boy has N rupees and he wants to get as many chocolates for free. The values of X, Y, M and N are passed as the input. The program must print the maximum number of chocolates he can buy from N rupees as the output.